package 重构
Former-commit-id: a04db4883ef26febc6855e66ec806303f5935785

+++ mynlp-segment/src/main/java/com/mayabot/nlp/segment/wordnet/VertexRow.java
@@ -52,7 +52,7 @@ public class VertexRow implements Iterable<Vertex> {
         this.wordnet = wordnet;
     }
 
-    public Vertex getOrCrete(short length) {
+    public Vertex getOrCrete(int length) {
 
         Vertex v = get(length);
